Organize AdyenCheckout module folder structure
