+++ AdyenCheckout/Callbacks/CheckoutResultCallbackStore.swift
