+++ AdyenCheckout/Flows/SessionCheckout.swift
@@ -1,0 +1,19 @@
+//
+// Copyright (c) 2026 Adyen N.V.
+//
+// This file is open source and available under the MIT license. See the LICENSE file for more info.
+//
+
+import Adyen
+
+/// Checkout flow for integrations using the `/sessions` endpoint.
+@MainActor
+public final class SessionCheckout: PaymentCheckout {
+
+    package let callbackStore: SessionCheckoutCallbackStore
+
+    package init(core: CheckoutCoreProtocol, callbackStore: SessionCheckoutCallbackStore) {
+        self.callbackStore = callbackStore
+        super.init(core: core, resultCallbacks: callbackStore)
+    }
+}
