+++ AdyenCheckout/Setup/CheckoutProviderProtocol.swift
